requirements.txt
pycairo==1.20.1
urllib3==1.26.8
requests==2.27.1
